api/requirements.txt
PyJWT>=2.0.0 